requirements.txt
fastapi
uvicorn
pydantic
cutlet
slowapi
unidic
python-dotenv
fastapi[standard]
worldnewsapi
deepl